requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
MySQL-python==1.2.5
pytz==2018.7
requests==2.20.1
schema==0.6.8
Schemer==0.2.11
shortuuid==0.5.0
SQLAlchemy==1.2.14
urllib3==1.24.1
Werkzeug==0.14.1
whenpy==0.3.0
